Created BasePlantPacket.cs
Big commit here, I'll try to document everything I can.

First up: Refactored the Sun.cs file, which is the custom resource manager.
  This involved:
    - Renaming some fields, namely SunCurrent -> CurrentSun
    - Adding some new fields, namely CurrentSunDisplay (used for making the ui number scale smoothly but have the actual sun value change instantly), SunDifference to facilitate this, and MaxSun2 as a buffer variable, among some others.
      - Other important ones are used for resource usage decreasing, one is multiplicative, the other is additive
  ResetVariables() (called in ResetEffects() and UpdateDead() ) has been changed to accommodate these new variables and renamings
  UpdateResource() has also been changed, now uses CurrentSun over old RegenAmount, and changed magic number 25 to DefaultRegenAmount
    - The UI smoothing has been updated to work when both increasing and decreasing:
      - Gets difference between the UI value and the internal value
      - Uses Math.Abs to check the difference/25 (arbitrary number), increments the UI value by that number.
      - Slightly faster increments for lower differences

Renamed damage class "Plants" to "PlantDamage", but changed the localized name to "Summon Damage", the class inherits all values associated with summon, but remains to be used for the UI (which shows when the player holds an item from the class)

Created "BasePlantPacket" and moved many common functions from Utilities into it.
  - Changed all the other packets from ": ModItem" to : BasePlantPacket"

Made localisation for the cooldown and sun usage of the plants.

Added a null value to PlantID.cs and shifted all the packets up by 1

Removed the QuickItem class from utilities, and updated rotobaga to reflect this

I've probably missed something but if I can't remember it it probably wasn't important.
This should've been like 4 commits lol.